+++ src/resfinder/run_resfinder.py
@@ -22,9 +22,6 @@ from resfinder.cge.phenotype2genotype.res_sumtable import PanelNameError
 
 from resfinder import __version__
 
-# TODO list:
-# TODO: JSON output summary is not species dependent
-
 
 def eprint(*args, **kwargs):
     print(*args, file=sys.stderr, **kwargs)
@@ -100,9 +97,10 @@ def main():
                         help="Input fasta file.",
                         default=None)
     parser.add_argument("-ifq", "--inputfastq",
-                        help="Input fastq file(s). Assumed to be single-end fastq \
-                              if only one file is provided, and assumed to be \
-                              paired-end data if two files are provided.",
+                        help=("Input fastq file(s). Assumed to be single-end "
+                              "fastq if only one file is provided, and assumed"
+                              " to be paired-end data if two files are "
+                              "provided."),
                         nargs="+",
                         default=None)
     parser.add_argument("--nanopore",
@@ -111,13 +109,14 @@ def main():
                         help="If nanopore data is used",
                         default=False)
     parser.add_argument("-o", "--outputPath",
-                        help=("Output directory. If it doesn't exist, it will be "
-                              "created."),
+                        help=("Output directory. If it doesn't exist, it will "
+                              "be created."),
                         required=True,
                         default=None)
     parser.add_argument("-j", "--out_json",
-                        help=("Specify JSON filename and output directory. If the "
-                              "directory doesn't exist, it will be created."),
+                        help=("Specify JSON filename and output directory. If "
+                              "the directory doesn't exist, it will be "
+                              "created."),
                         default=None)
     parser.add_argument("-b", "--blastPath",
                         help="Path to blastn",
@@ -130,22 +129,21 @@ def main():
                         default=None)
     parser.add_argument("--ignore_missing_species",
                         action="store_true",
-                        help="If set, species is provided and --point flag is set, "
-                             "will not throw an error if no database is found for "
-                             "the provided species. If species is not found. Point "
-                             "mutations will silently be ignored.",
+                        help=("If set, species is provided and --point flag "
+                              "is set, will not throw an error if no database "
+                              "is found for the provided species. If species "
+                              "is not found. Point mutations will silently "
+                              "be ignored."),
                         default=False)
-
 
     # Acquired resistance options
     parser.add_argument("-db_res", "--db_path_res",
-                        help=("Path to the databases for ResFinder. Defaults to "
-                              "'db_resfinder' in the ResFinder application "
-                              "directory."),
+                        help=("Path to the databases for ResFinder."),
                         default=None)
     parser.add_argument("-db_res_kma", "--db_path_res_kma",
-                        help=("Path to the ResFinder databases indexed with KMA. "
-                              "Defaults to the value of the --db_res flag."),
+                        help=("Path to the ResFinder databases indexed with "
+                              "KMA. Defaults to the value of the --db_res "
+                              "flag."),
                         default=None)
     parser.add_argument("-acq", "--acquired",
                         action="store_true",
@@ -158,8 +156,8 @@ def main():
                         type=int,
                         default=None)
     parser.add_argument("-l", "--min_cov",
-                        help=("Minimum (breadth-of) coverage of ResFinder within "
-                              "the range 0-1."),
+                        help=("Minimum (breadth-of) coverage of ResFinder "
+                              "within the range 0-1."),
                         type=float,
                         default=None)
     parser.add_argument("-t", "--threshold",
@@ -190,9 +188,9 @@ def main():
                         help="Path to the databases for PointFinder",
                         default=None)
     parser.add_argument("-db_point_kma", "--db_path_point_kma",
-                        help="Path to the PointFinder databases indexed with KMA. \
-                              Defaults to the 'kma_indexing' directory inside the \
-                              given database directory.",
+                        help=("Path to the PointFinder databases indexed with "
+                              "KMA. Defaults to the value of the "
+                              "--db_path_point flag."),
                         default=None)
     parser.add_argument("-g", "--specific_gene",
                         nargs='+',
@@ -202,24 +200,25 @@ def main():
                         default=None)
     parser.add_argument("-u", "--unknown_mut",
                         action="store_true",
-                        help="Show all mutations found even if in unknown to the\
-                              resistance database",
+                        help=("Show all mutations found even if in unknown to "
+                              "the resistance database"),
                         default=None)
     parser.add_argument("-l_p", "--min_cov_point",
-                        help=("Minimum (breadth-of) coverage of Pointfinder within "
-                              "the range 0-1. If None is selected, the minimum "
-                              "coverage of ResFinder will be used."),
+                        help=("Minimum (breadth-of) coverage of Pointfinder "
+                              "within the range 0-1. If None is selected, the "
+                              "minimum coverage of ResFinder will be used."),
                         type=float,
                         default=None)
     parser.add_argument("-t_p", "--threshold_point",
-                        help=("Threshold for identity of Pointfinder within the "
-                              "range 0-1. If None is selected, the minimum "
-                              "coverage of ResFinder will be used."),
+                        help=("Threshold for identity of Pointfinder within "
+                              "the range 0-1. If None is selected, the minimum"
+                              " coverage of ResFinder will be used."),
                         type=float,
                         default=None)
     parser.add_argument("--ignore_indels",
                         action="store_true",
-                        help="Ignore frameshift-causing indels in Pointfinder.",
+                        help=("Ignore frameshift-causing indels in "
+                              "Pointfinder."),
                         default=None)
     parser.add_argument("--ignore_stop_codons",
                         action="store_true",
@@ -232,9 +231,9 @@ def main():
     # Temporary option only available temporary
     parser.add_argument("--pickle",
                         action="store_true",
-                        help="Create a pickle dump of the Isolate object. \
-                              Currently needed in the CGE webserver. Dependency \
-                              and this option is being removed.",
+                        help=("Create a pickle dump of the Isolate object. "
+                              "Currently needed in the CGE webserver. "
+                              "Dependency and this option is being removed."),
                         default=False)
 
     args = parser.parse_args()
@@ -262,6 +261,17 @@ def main():
         std_result.init_database("DisinFinder", conf.db_path_disinf)
 
     kma_resultfiles = ["Result", "Fragments"]
+
+    # Load genotype to phenotype database
+    res_pheno_db = PhenoDB(
+        abclassdef_file=conf.abclassdef_file,
+        acquired_file=conf.phenotype_file,
+        point_file=conf.point_file,
+        disinf_file=conf.disinf_file,
+        disclassdef_file=conf.disclassdef_file
+    )
+
+
     ##########################################################################
     # ResFinder
     ##########################################################################
@@ -275,6 +285,7 @@ def main():
                                     db_path_kma=conf.db_path_res_kma)
 
         if(conf.inputfasta):
+
             method = ResFinder.TYPE_BLAST
             resfinder_result = acquired_finder.blast(
                 inputfile=conf.inputfasta,
@@ -282,10 +293,24 @@ def main():
                 min_cov=conf.rf_gene_cov,
                 threshold=conf.rf_gene_id,
                 blast=conf.blast,
-                allowed_overlap=conf.rf_overlap)
+                allowed_overlap=conf.rf_overlap
+            )
+
+            # DEPRECATED
+            # TODO: make a write method that depends on the json output
+            acquired_finder.write_results(out_path=conf.outputPath,
+                                          result=blast_results,
+                                          res_type=ResFinder.TYPE_BLAST,
+                                          software="ResFinder")
+
+            ResFinderResultHandler.standardize_results(std_result,
+                                                       blast_results.results,
+                                                       "ResFinder")
+
         else:
             method = ResFinder.TYPE_KMA
             if(conf.nanopore):
+
                 kma_params = dict(inputfile_1=conf.inputfastq_1,
                                   inputfile_2=conf.inputfastq_2,
                                   output=conf.outPath_res_kma,
@@ -337,6 +362,7 @@ def main():
         else:
             method = ResFinder.TYPE_KMA
             if(conf.nanopore):
+
                 kma_params = dict(inputfile_1=conf.inputfastq_1,
                                   inputfile_2=conf.inputfastq_2,
                                   output=conf.outPath_disinf_kma,
@@ -414,17 +440,14 @@ def main():
             results = kma_manager.run_KMAAligner(conf, kma_resultfiles,
                                                  kma_params)
 
+
         handle_results(finder=finder, res=results, std_result=std_result,
                        conf=conf, db_name="PointFinder", method=method)
 
     ##########################################################################
     # Phenotype to genotype
     ##########################################################################
-    # Load genotype to phenotype database
-    res_pheno_db = PhenoDB(
-        abclassdef_file=conf.abclassdef_file, acquired_file=conf.phenotype_file,
-        point_file=conf.point_file, disinf_file=conf.disinf_file,
-        disclassdef_file=conf.disclassdef_file)
+
     # Isolate object store results
     isolate = Isolate(name=conf.sample_name, species=conf.species,
                       amr_panel_file=conf.db_panels_file)